chore: merge RowOperation into mod.rs
